Update playtime collection to include Echo
Echo was released between week 3 and 4 of play. Due to her release, post-week-3 there was no collection of Sigma playtime. To account for this the playtime for Sigma does not include weeks 4 through 6. Echo playtime will include week 5 even though she was out of the pool, but not week 4 and she was not in competitive play.